requirements.txt
uvicorn[standard]>=0.30
gunicorn>=22.0
django
python-jose[cryptography]
PyJWT